Add backwards compatibility for replicate capability.
Most test cases have cells that are smaller than the model cutoff,
and replicating them is invalid since they are liquid. Introduced
a variable (allow_replication) that can be set during when the
serial_calculator is instantiated, and defaults to false.

NOT updated yet for any of the other languages!